# internal/handler/http/gen/openapi.gen.go
// Package gen provides primitives to interact the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package gen

import (
	"fmt"
	"net/http"

	"github.com/deepmap/oapi-codegen/pkg/runtime"
	"github.com/go-chi/chi"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Gets converted currency
	// (GET /v1/conversions)
	GetConversionAmount(w http.ResponseWriter, r *http.Request, params GetConversionAmountParams)
	// Creates a new conversion rate
	// (POST /v1/conversions)
	CreateConversion(w http.ResponseWriter, r *http.Request, params CreateConversionParams)
	// Returns all currencies stored in database
	// (GET /v1/currencies)
	GetCurrency(w http.ResponseWriter, r *http.Request, params GetCurrencyParams)
	// Creates a new currency
	// (POST /v1/currencies)
	CreateCurrency(w http.ResponseWriter, r *http.Request, params CreateCurrencyParams)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
}

type MiddlewareFunc func(http.HandlerFunc) http.HandlerFunc

// GetConversionAmount operation middleware
func (siw *ServerInterfaceWrapper) GetConversionAmount(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetConversionAmountParams

	// ------------- Required query parameter "currency_id_from" -------------
	if paramValue := r.URL.Query().Get("currency_id_from"); paramValue != "" {

	} else {
		http.Error(w, "Query argument currency_id_from is required, but not found", http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "currency_id_from", r.URL.Query(), &params.CurrencyIdFrom)
	if err != nil {
		http.Error(w, fmt.Sprintf("Invalid format for parameter currency_id_from: %s", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "currency_id_to" -------------
	if paramValue := r.URL.Query().Get("currency_id_to"); paramValue != "" {

	} else {
		http.Error(w, "Query argument currency_id_to is required, but not found", http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "currency_id_to", r.URL.Query(), &params.CurrencyIdTo)
	if err != nil {
		http.Error(w, fmt.Sprintf("Invalid format for parameter currency_id_to: %s", err), http.StatusBadRequest)
		return
	}

	// ------------- Required query parameter "amount" -------------
	if paramValue := r.URL.Query().Get("amount"); paramValue != "" {

	} else {
		http.Error(w, "Query argument amount is required, but not found", http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "amount", r.URL.Query(), &params.Amount)
	if err != nil {
		http.Error(w, fmt.Sprintf("Invalid format for parameter amount: %s", err), http.StatusBadRequest)
		return
	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetConversionAmount(w, r, params)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// CreateConversion operation middleware
func (siw *ServerInterfaceWrapper) CreateConversion(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params CreateConversionParams

	headers := r.Header

	// ------------- Optional header parameter "Authorization" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Authorization")]; found {
		var Authorization string
		n := len(valueList)
		if n != 1 {
			http.Error(w, fmt.Sprintf("Expected one value for Authorization, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameter("simple", false, "Authorization", valueList[0], &Authorization)
		if err != nil {
			http.Error(w, fmt.Sprintf("Invalid format for parameter Authorization: %s", err), http.StatusBadRequest)
			return
		}

		params.Authorization = &Authorization

	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateConversion(w, r, params)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// GetCurrency operation middleware
func (siw *ServerInterfaceWrapper) GetCurrency(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetCurrencyParams

	headers := r.Header

	// ------------- Optional header parameter "Authorization" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Authorization")]; found {
		var Authorization string
		n := len(valueList)
		if n != 1 {
			http.Error(w, fmt.Sprintf("Expected one value for Authorization, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameter("simple", false, "Authorization", valueList[0], &Authorization)
		if err != nil {
			http.Error(w, fmt.Sprintf("Invalid format for parameter Authorization: %s", err), http.StatusBadRequest)
			return
		}

		params.Authorization = &Authorization

	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetCurrency(w, r, params)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// CreateCurrency operation middleware
func (siw *ServerInterfaceWrapper) CreateCurrency(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params CreateCurrencyParams

	headers := r.Header

	// ------------- Optional header parameter "Authorization" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Authorization")]; found {
		var Authorization string
		n := len(valueList)
		if n != 1 {
			http.Error(w, fmt.Sprintf("Expected one value for Authorization, got %d", n), http.StatusBadRequest)
			return
		}

		err = runtime.BindStyledParameter("simple", false, "Authorization", valueList[0], &Authorization)
		if err != nil {
			http.Error(w, fmt.Sprintf("Invalid format for parameter Authorization: %s", err), http.StatusBadRequest)
			return
		}

		params.Authorization = &Authorization

	}

	var handler = func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateCurrency(w, r, params)
	}

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler(w, r.WithContext(ctx))
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL     string
	BaseRouter  chi.Router
	Middlewares []MiddlewareFunc
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/conversions", wrapper.GetConversionAmount)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/conversions", wrapper.CreateConversion)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/v1/currencies", wrapper.GetCurrency)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/v1/currencies", wrapper.CreateCurrency)
	})

	return r
}