# internal/handler/http/gen/openapi_types.gen.go
// Package gen provides primitives to interact the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package gen

// ConversionRequest defines model for ConversionRequest.
type ConversionRequest struct {
	CurrencyIdFrom int64 `json:"currency_id_from"`
	CurrencyIdTo   int64 `json:"currency_id_to"`
	Rate           int64 `json:"rate"`
}

// ConversionReturn defines model for ConversionReturn.
type ConversionReturn struct {
	Error  *string `json:"error,omitempty"`
	Result int64   `json:"result"`
}

// CreateCurrencyInvalidPayload defines model for CreateCurrencyInvalidPayload.
type CreateCurrencyInvalidPayload struct {
	Detail *struct {
		Name *string `json:"name,omitempty"`
	} `json:"detail,omitempty"`
	Error *string `json:"error,omitempty"`
}

// CreateSuccessReturn defines model for CreateSuccessReturn.
type CreateSuccessReturn struct {
	Status *string `json:"status,omitempty"`
}

// Currency defines model for Currency.
type Currency struct {
	Id   int64  `json:"id"`
	Name string `json:"name"`
}

// CurrencyList defines model for CurrencyList.
type CurrencyList struct {
	Data []Currency `json:"data"`
}

// GenericError defines model for GenericError.
type GenericError struct {
	Error *string `json:"error,omitempty"`
}

// GetConversionAmountParams defines parameters for GetConversionAmount.
type GetConversionAmountParams struct {

	// source currency id to get amount
	CurrencyIdFrom int64 `json:"currency_id_from"`

	// target currency id to get amount
	CurrencyIdTo int64 `json:"currency_id_to"`

	// amount to get converted
	Amount int64 `json:"amount"`
}

// CreateConversionJSONBody defines parameters for CreateConversion.
type CreateConversionJSONBody ConversionRequest

// CreateConversionParams defines parameters for CreateConversion.
type CreateConversionParams struct {

	// Base64 encoded squad / tribe name
	Authorization *string `json:"Authorization,omitempty"`
}

// GetCurrencyParams defines parameters for GetCurrency.
type GetCurrencyParams struct {
	Authorization *string `json:"Authorization,omitempty"`
}

// CreateCurrencyJSONBody defines parameters for CreateCurrency.
type CreateCurrencyJSONBody Currency

// CreateCurrencyParams defines parameters for CreateCurrency.
type CreateCurrencyParams struct {
	Authorization *string `json:"Authorization,omitempty"`
}

// CreateConversionJSONRequestBody defines body for CreateConversion for application/json ContentType.
type CreateConversionJSONRequestBody CreateConversionJSONBody

// CreateCurrencyJSONRequestBody defines body for CreateCurrency for application/json ContentType.
type CreateCurrencyJSONRequestBody CreateCurrencyJSONBody